requirements.txt
streamlit
click
ruff
PyPDF2
pillow
pytesseract
langchain
langchain_openai
langchain-community langchain-openai pydantic
